feat(tasks): add bulk quick-add creation to the task store
createNewTasksBulk builds full task payloads from quick add magic
titles (extracted from createNewTask so both paths share it) and hands
them to the service in one call. Label application failures toast and
continue — the tasks already exist server-side, failing the whole
action would invite duplicate resubmits.

runWrites moves to helpers/ so components can use the write-serializing
util without importing the store module.
