+++ frontend/src/stores/tasks.test.ts
@@ -17,7 +17,7 @@ vi.mock('@/stores/base', () => ({
 	useBaseStore: () => ({setHasTasks: vi.fn()}),
 }))
 
-import {buildDefaultRemindersForQuickAdd, runWrites, useTaskStore} from './tasks'
+import {buildDefaultRemindersForQuickAdd, useTaskStore} from './tasks'
 import {useLabelStore} from './labels'
 import LabelModel from '@/models/label'
 import {REMINDER_PERIOD_RELATIVE_TO_TYPES} from '@/types/IReminderPeriodRelativeTo'
@@ -61,42 +61,6 @@ describe('buildDefaultRemindersForQuickAdd', () => {
 		const weird = {...aDefault, relativeTo: REMINDER_PERIOD_RELATIVE_TO_TYPES.STARTDATE} as ITaskReminder
 		const result = buildDefaultRemindersForQuickAdd([weird], '2026-05-01T00:00:00.000Z')
 		expect(result[0].relativeTo).toBe(REMINDER_PERIOD_RELATIVE_TO_TYPES.DUEDATE)
-	})
-})
-
-describe('runWrites', () => {
-	function deferredWrite() {
-		const inFlight: string[] = []
-		let maxConcurrent = 0
-		const completed: string[] = []
-		const write = async (item: string) => {
-			inFlight.push(item)
-			maxConcurrent = Math.max(maxConcurrent, inFlight.length)
-			await Promise.resolve()
-			inFlight.splice(inFlight.indexOf(item), 1)
-			completed.push(item)
-		}
-		return {write, completed, getMaxConcurrent: () => maxConcurrent}
-	}
-
-	it('runs all writes in parallel when concurrent', async () => {
-		const {write, completed, getMaxConcurrent} = deferredWrite()
-		await runWrites(['a', 'b', 'c'], write, true)
-		expect(completed).toHaveLength(3)
-		expect(getMaxConcurrent()).toBeGreaterThan(1)
-	})
-
-	it('runs writes one at a time when not concurrent', async () => {
-		const {write, completed, getMaxConcurrent} = deferredWrite()
-		await runWrites(['a', 'b', 'c'], write, false)
-		expect(completed).toEqual(['a', 'b', 'c'])
-		expect(getMaxConcurrent()).toBe(1)
-	})
-
-	it('does nothing for an empty list', async () => {
-		const {write, completed} = deferredWrite()
-		await runWrites([], write, false)
-		expect(completed).toHaveLength(0)
 	})
 })
